moved BFH library
