+++ src/main/java/com/cburch/logisim/std/bfh/bin2bcd/add3.fig
